requirements.txt
# SET Financial — EOM Close Automation & QBO MCP Server
# Install: pip install -r requirements.txt

# Core data processing
pandas>=2.0
openpyxl>=3.1          # Excel read/write
pdfplumber>=0.10       # PDF parsing (DPV, Pier statements)
python-dotenv>=1.0     # .env loading

# QuickBooks Online API
requests>=2.31

# MCP (Model Context Protocol) server
mcp>=1.0

# Dev / testing
pytest>=7.4
tabulate>=0.9          # DataFrame.to_markdown()
